Organize logging: file-based logs for master/workers, rename logging.py to log.py
- Add add_file_handler() to attach per-process log files dynamically
- Master process logs to {outdir}/logs/master.log
- Worker processes log to {outdir}/logs/worker_{rank}.log
- Rename walker logs to worker_salt2mu_{rank}.log for clarity
- Fix MPI worker detection bug in mcmc.py (config.USE_MPI on None)
- Rename logging.py to log.py to avoid shadowing stdlib logging
- Remove unused parallel/ output directory
- Update README with output directory tree
